Rewrite the text extraction functions around pdfium API limitations
The pdfium getters for text rects is counter-intuitive : when you request
the number of text areas for a range of characters, it will store the
rectangles, altering future calls to GetRect.
So the best thing is not not expose these dangerous calls and just
forward the vector of text rects.